DONE - task 24: Proxies should be made with EMF Proxy and not javassist http://code.google.com/a/eclipselabs.org/p/emftriple/issues/detail?id=24
UPDATE - add preliminary support for multiple data sources.

UPDATE - moved dependencies to Xtext in another plugins. This is done to allow usage of emftriple as a 
server side data mapper for EMF on GWT. Xtext is not supported in current EMF4GWT implementation (URIHandler issue??). As a 
result JPQL is not supported in emftriple on GWT.

ADDED - com.emftriple.query is a new plugin that contains JPQL, SPARQL engines (rely on Xtext)